ap dung tinh nang dang nhap

+++ src/main/java/com/codegym/controller/BlogServlet.java
@@ -26,8 +26,8 @@ public class BlogServlet extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HttpSession session = request.getSession();
-        User user = (User) session.getAttribute("user");
-        if (user == null) {
+        User loginUser = (User) session.getAttribute("user");
+        if (loginUser == null) {
             response.sendRedirect("");
         }
         else {
@@ -91,38 +91,46 @@ public class BlogServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String action = request.getParameter("action");
-        if (action == null) {
-            action = "";
+        HttpSession session = request.getSession();
+        User loginUser = (User) session.getAttribute("user");
+        if (loginUser == null) {
+            response.sendRedirect("");
         }
-        switch (action) {
-            case "create": {
-                int category_id = Integer.parseInt(request.getParameter("category_id"));
-                int user_id = 1;
-                String tittle = request.getParameter("tittle");
-                String content = request.getParameter("content");
-                Blog blog = new Blog(category_id, user_id, tittle, content);
-                blogService.create(blog);
-                response.sendRedirect("/blogs");
-                break;
+        else {
+            String action = request.getParameter("action");
+            if (action == null) {
+                action = "";
             }
-            case "edit": {
-                int id = Integer.parseInt(request.getParameter("id"));
-                int user_id = 1;
-                int category_id = Integer.parseInt(request.getParameter("category_id"));
-                String tittle = request.getParameter("tittle");
-                String content = request.getParameter("content");
-                Blog blog = new Blog(id, category_id, user_id, tittle, content);
-                blogService.updateById(id, blog);
-                response.sendRedirect("/blogs");
-                break;
-            }
-            case "delete": {
-                int id = Integer.parseInt(request.getParameter("id"));
-                blogService.deleteById(id);
-                response.sendRedirect("/blogs");
-                break;
+            switch (action) {
+                case "create": {
+                    int category_id = Integer.parseInt(request.getParameter("category_id"));
+                    int user_id =loginUser.getId();
+                    String tittle = request.getParameter("tittle");
+                    String content = request.getParameter("content");
+                    Blog blog = new Blog(category_id, user_id, tittle, content);
+                    blogService.create(blog);
+                    response.sendRedirect("/blogs");
+                    break;
+                }
+                case "edit": {
+                    int id = Integer.parseInt(request.getParameter("id"));
+                    int user_id = 1;
+                    int category_id = Integer.parseInt(request.getParameter("category_id"));
+                    String tittle = request.getParameter("tittle");
+                    String content = request.getParameter("content");
+                    Blog blog = new Blog(id, category_id, user_id, tittle, content);
+                    blogService.updateById(id, blog);
+                    response.sendRedirect("/blogs");
+                    break;
+                }
+                case "delete": {
+                    int id = Integer.parseInt(request.getParameter("id"));
+                    blogService.deleteById(id);
+                    response.sendRedirect("/blogs");
+                    break;
+                }
             }
         }
+
     }
 }
